+++ renderer/lib/hx.js
@@ -1,0 +1,5 @@
+var h = require('virtual-dom/h')
+var hyperx = require('hyperx')
+var hx = hyperx(h)
+
+module.exports = hx
