+++ renderer/main.html
@@ -3,9 +3,9 @@
   <head>
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel="stylesheet" href="index.css" charset="utf-8">
+    <link rel="stylesheet" href="main.css" charset="utf-8">
   </head>
   <body>
-    <script async src="index.js"></script>
+    <script async src="main.js"></script>
   </body>
 </html>
